Delete all station when less than 6xx And Refactor in Main activity

+++ app/src/main/java/io/github/shredktp/trainschedulesrt/data/source/pair_station/PairStationDataSource.java
@@ -17,5 +17,5 @@ public interface PairStationDataSource {
     PairStation getSeeFirstPairStation() throws Exception;
     ArrayList<PairStation> getAllPairStation();
 
-    long deleteAll();
+    int deleteAll();
 }
